feat: separate context utils
