+++ src/index.ts
@@ -3,5 +3,6 @@ import Command from './command';
 import ApiHelper from './api-helper';
 import GitHelper from './git-helper';
 import * as Utils from './utils';
+import * as ContextHelper from './context-helper';
 
-export { Logger, Command, ApiHelper, GitHelper, Utils };
+export { Logger, Command, ApiHelper, GitHelper, Utils, ContextHelper };
